+++ app/frontend/lib/extensions/suggestion/view.svelte
@@ -1,15 +1,11 @@
 <script lang="ts">
   import { cn } from '@/lib/utils'
-  import type { Editor } from '@tiptap/core'
-  import type { Node } from '@tiptap/pm/model'
+  import type { NodeViewProps } from '@tiptap/core'
   import { NodeViewContent, NodeViewWrapper } from 'svelte-tiptap'
-  import { twMerge } from 'tailwind-merge'
 
-  export let node: Node
-  export let editor: Editor
-  export let getPos
+  let { node, editor }: NodeViewProps = $props()
 
-  let isHovered = false
+  let isHovered = $state(false)
 
   function acceptSuggestion() {
     const suggestionId = node.attrs['data-id']
@@ -59,19 +55,17 @@
   {...node.attrs}
   class={cn(
     'relative data-[action=add]:bg-green-900 data-[action=remove]:bg-red-900 data-[action]:px-0.5 rounded data-action:border-2',
-    node.attrs['data-empty'] ? 'hidden' : '',
-    node.attrs['data-empty-brother']
-      ? 'my-5'
-      : node.attrs['data-action'] === 'add'
-        ? 'mb-5'
-        : 'mt-5'
+    { 'hidden': node.attrs['data-empty'] },
+    { 'my-5': node.attrs['data-empty-brother'] },
+    { 'mb-5': node.attrs['data-action'] === 'add' },
+    { 'mt-5': node.attrs['data-action'] === 'remove' }
   )}
   onmouseenter={() => (isHovered = true)}
   onmouseleave={() => (isHovered = false)}
 >
   {#if !!node.attrs['data-action'] && isHovered}
     <div
-      class={twMerge(
+      class={cn(
         'absolute z-10 left-0 flex gap-1 border-b-white/5',
         node.attrs['data-action'] === 'add'
           ? '-bottom-7 border-b-2'
@@ -79,14 +73,14 @@
       )}
     >
       <button
-        on:click={acceptSuggestion}
+        onclick={acceptSuggestion}
         contenteditable="false"
         class="text-green-500 px-1 cursor-pointer transition hover:bg-white/5"
       >
         Aceitar
       </button>
       <button
-        on:click={refuseSuggestion}
+        onclick={refuseSuggestion}
         contenteditable="false"
         class="text-red-500 px-1 cursor-pointer transition hover:bg-white/5"
       >
@@ -94,7 +88,7 @@
       </button>
       {#if !node.attrs['data-empty'] && !node.attrs['data-empty-brother']}
         <button
-          on:click={keepBothFromSuggestion}
+          onclick={keepBothFromSuggestion}
           contenteditable="false"
           class="text-orange-400 px-1 cursor-pointer transition hover:bg-white/5"
         >
