+++ app/frontend/services/ai-api-service.ts
@@ -1,7 +1,9 @@
 import axios from 'axios'
 
+const url = import.meta.env.VITE_AI_SERVICE_URL
+
 export const aiApiService = axios.create({
-  baseURL: 'http://localhost:8000',
+  baseURL: url,
   headers: {
     'Content-Type': 'application/json',
   },
